Task4 - AZURE LANGUAGE SERVICES/KeyPhraseExtraction - Task4/requirement.txt
azure-ai-textanalytics
azure-core==1.23.0
Flask==2.0.1
Flask